Statistics.Distribution.Dirichlet

Description

Creation date: Tue Oct 20 10:10:39 2020.

Dirichlet distribution

data DirichletDistribution

The Dirichlet distribution.

dirichletDistribution :: Vector Double -> Either String DirichletDistribution

Create a Dirichlet distribution from the given parameter vector.

Return Left if:

The parameter vector has less then two elements.

One or more parameters are negative or zero.

dirichletDensity :: DirichletDistribution -> Vector Double -> Log Double

Density of the Dirichlet distribution evaluated at a given value vector.

Return 0 if:

The value vector has a different length than the parameter vector.

The value vector has elements being negative or zero.

The value vector does not sum to 1.0 (with tolerance eps = 1e-14).

dirichletSample :: StatefulGen g m => DirichletDistribution -> g -> m (Vector Double)

Sample a value vector from the Dirichlet distribution.

Symmetric Dirichlet distribution

data DirichletDistributionSymmetric

See DirichletDistribution but with parameter vector replicate DIM VAL.

dirichletDistributionSymmetric :: Int -> Double -> Either String DirichletDistributionSymmetric

Create a symmetric Dirichlet distribution of given dimension and parameter.

Return Left if:

The given dimension is smaller than two.

The parameter is negative or zero.

dirichletDensitySymmetric :: DirichletDistributionSymmetric -> Vector Double -> Log Double

Density of the symmetric Dirichlet distribution evaluated at a given value vector.

Return 0 if:

The value vector has a different dimension.

The value vector has elements being negative or zero.

The value vector does not sum to 1.0 (with tolerance eps = 1e-14).

dirichletSampleSymmetric :: StatefulGen g m => DirichletDistributionSymmetric -> g -> m (Vector Double)

Sample a value vector from the symmetric Dirichlet distribution.